+++ ui/user/src/routes/admin/device-mcp-servers/[hash]/+page.ts
@@ -2,5 +2,5 @@ import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
 export const load: PageLoad = ({ params, url }) => {
-	throw redirect(301, `/admin/devices/mcp-servers/${encodeURIComponent(params.hash)}${url.search}`);
+	throw redirect(301, `/inventory/mcp-servers/${encodeURIComponent(params.hash)}${url.search}`);
 };
