+++ ui/user/src/routes/admin/device-skills/+page.ts
@@ -4,5 +4,5 @@ import { redirect } from '@sveltejs/kit';
 export const load: PageLoad = ({ url }) => {
 	const searchParams = new URLSearchParams(url.searchParams);
 	searchParams.set('view', 'device-skills');
-	throw redirect(301, `/admin/devices?${searchParams}`);
+	throw redirect(301, `/inventory?${searchParams}`);
 };
