+++ ui/user/src/routes/admin/device-skills/[name]/+page.ts
@@ -2,5 +2,5 @@ import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
 export const load: PageLoad = ({ params, url }) => {
-	throw redirect(301, `/admin/devices/skills/${encodeURIComponent(params.name)}${url.search}`);
+	throw redirect(301, `/inventory/skills/${encodeURIComponent(params.name)}${url.search}`);
 };
