+++ ui/user/src/routes/admin/filters/c/[filterEntryId]/+page.ts
@@ -2,5 +2,8 @@ import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
 export const load: PageLoad = ({ params, url }) => {
-	throw redirect(301, `/mcp-servers/access-policies/${encodeURIComponent(params.id)}${url.search}`);
+	throw redirect(
+		301,
+		`/mcp-servers/filters/c/${encodeURIComponent(params.filterEntryId)}${url.search}`
+	);
 };
