--- ui/user/src/routes/admin/hosted-agent-access-policies/+layout.ts
@@ -1,7 +1,0 @@
-import { requireHostedAgentsEnabled } from '$lib/hostedAgents';
-import type { LayoutLoad } from './$types';
-
-export const load: LayoutLoad = async ({ parent }) => {
-	const { version } = await parent();
-	requireHostedAgentsEnabled(version, '/dashboard');
-};
