+++ ui/user/src/routes/admin/model-access-policies/[id]/+page.ts
@@ -1,19 +1,6 @@
-import { handleRouteError } from '$lib/errors';
-import { AdminService } from '$lib/services';
-import { profile } from '$lib/stores';
 import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async ({ params, fetch }) => {
-	const { id } = params;
-
-	let modelAccessPolicy;
-	try {
-		modelAccessPolicy = await AdminService.getModelAccessPolicy(id, { fetch });
-	} catch (err) {
-		handleRouteError(err, `/admin/model-access-policies/${id}`, profile.current);
-	}
-
-	return {
-		modelAccessPolicy
-	};
+export const load: PageLoad = ({ params, url }) => {
+	throw redirect(301, `/models/access-policies/${encodeURIComponent(params.id)}${url.search}`);
 };
