+++ ui/user/src/routes/admin/skill-access-policies/+page.ts
@@ -4,5 +4,5 @@ import { redirect } from '@sveltejs/kit';
 export const load: PageLoad = ({ url }) => {
 	const searchParams = new URLSearchParams(url.searchParams);
 	searchParams.delete('view');
-	throw redirect(301, `/skills?view=policies&${searchParams}`);
+	throw redirect(301, `/skills?view=access-policies&${searchParams}`);
 };
