+++ ui/user/src/routes/inventory/devices/[device_id]/+page.svelte
