+++ ui/user/src/routes/inventory/devices/[device_id]/scans/[scan_id]/+layout.ts
