+++ ui/user/src/routes/inventory/devices/[device_id]/scans/[scan_id]/+page.svelte
