+++ ui/user/src/routes/inventory/devices/[device_id]/scans/[scan_id]/_shared/files.ts
