+++ ui/user/src/routes/inventory/devices/[device_id]/scans/[scan_id]/skills/[id]/+page.svelte
