+++ ui/user/src/routes/inventory/skills/[name]/+page.svelte
@@ -166,7 +166,7 @@
 					]}
 					onClickRow={(d, isCtrlClick) => {
 						openUrl(
-							resolve(`/admin/devices/${d.deviceID}/scans/${d.deviceScanID}/skills/${d.id}`),
+							resolve(`/inventory/devices/${d.deviceID}/scans/${d.deviceScanID}/skills/${d.id}`),
 							isCtrlClick
 						);
 					}}
@@ -174,7 +174,7 @@
 					{#snippet onRenderColumn(property, d: Row)}
 						{#if property === 'shortDeviceID'}
 							<a
-								href={resolve(`/admin/devices/${d.deviceID}`)}
+								href={resolve(`/inventory/devices/${d.deviceID}`)}
 								class="btn-link text-primary"
 								title={d.deviceID}
 								onclick={(e) => e.stopPropagation()}
