+++ ui/user/src/routes/mcp-servers/access-policies/[id]/+page.svelte
