+++ ui/user/src/routes/mcp-servers/access-policies/[id]/+page.ts
@@ -1,0 +1,19 @@
+import { handleRouteError } from '$lib/errors';
+import { AdminService } from '$lib/services';
+import { profile } from '$lib/stores';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ params, fetch }) => {
+	const { id } = params;
+
+	let accessControlRule;
+	try {
+		accessControlRule = await AdminService.getAccessControlRule(id, { fetch });
+	} catch (err) {
+		handleRouteError(err, `/admin/mcp-access-policies/${id}`, profile.current);
+	}
+
+	return {
+		accessControlRule
+	};
+};
