+++ ui/user/src/routes/mcp-servers/access-policies/w/[wid]/r/[id]/+page.svelte
