+++ ui/user/src/routes/skills/access-policies/[id]/+page.svelte
